aggregates/coverage: Use relative path for json files
Currently the aggregates and coverage use the download location path in
the datagetter_metadata block from the datagetter. However we want to
run reports on data that may be from the datastore or another source
(such as when it has been copied to another location) so use relative
path instead of full absolute path of the data.

Fixes https://github.com/ThreeSixtyGiving/datastore/issues/54